requirements.txt
PyQt6>=6.10.0
pyqtgraph>=0.13.0
PyAudio>=0.2.14
numpy>=1.24.0
PyYAML>=6.0
